# backend/openapi_cache.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "链客宝API",
    "description": "链客宝后端API服务 — Premium Business Network and Entrepreneur Supply-Demand Matching Platform。\n\n提供认证、产品、订单、搜索、支付、充值、发票、对账、联系人管理、活动时间线、数据洞察、供需匹配、AI智能匹配引擎、推广员体系、管理后台、系统配置等完整业务能力。",
    "contact": {
      "name": "链客宝团队",
      "url": "https://www.go-aiport.com/",
      "email": "support@go-aiport.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://www.go-aiport.com/"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/search": {
      "get": {
        "tags": [
          "搜索"
        ],
        "summary": "Search Products",
        "description": "产品搜索\n\n支持多维度筛选、全文搜索和中英文混合搜索。已上架 (approved) 产品可见。\n\n参数说明:\n- q: 搜索关键词，匹配产品名称、描述、标签、品牌（无需完整匹配，智能分词）\n- category: 按分类精确筛选\n- region: 按产地/地区筛选（模糊匹配规格JSON中的\"产地\"字段）\n- min_price / max_price: 价格区间筛选\n- sort_by: 排序方式\n  - relevance（默认）: 按文本相关性降序\n  - price_asc: 按价格升序（同价按相关性）\n  - price_desc: 按价格降序（同价按相关性）\n  - newest: 按创建时间降序（最新优先）\n- page / page_size: 分页\n- highlight: 是否返回高亮片段",
        "operationId": "search_products_api_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "搜索关键词（产品名称/描述/标签/品牌）",
              "default": "",
              "title": "Q"
            },
            "description": "搜索关键词（产品名称/描述/标签/品牌）"
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "分类筛选（精确匹配）",
              "title": "Category"
            },
            "description": "分类筛选（精确匹配）"
          },
          {
            "name": "region",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "地区筛选（匹配规格中的产地）",
              "title": "Region"
            },
            "description": "地区筛选（匹配规格中的产地）"
          },
          {
            "name": "min_price",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 0,
              "description": "最低价格",
              "title": "Min Price"
            },
            "description": "最低价格"
          },
          {
            "name": "max_price",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 0,
              "description": "最高价格",
              "title": "Max Price"
            },
            "description": "最高价格"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "排序方式: relevance(相关性) / price_asc(价格升序) / price_desc(价格降序) / newest(最新)",
              "default": "relevance",
              "title": "Sort By"
            },
            "description": "排序方式: relevance(相关性) / price_asc(价格升序) / price_desc(价格降序) / newest(最新)"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "页码",
              "default": 1,
              "title": "Page"
            },
            "description": "页码"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "每页条数",
              "default": 20,
              "title": "Page Size"
            },
            "description": "每页条数"
          },
          {
            "name": "highlight",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "是否返回高亮片段",
              "default": true,
              "title": "Highlight"
            },
            "description": "是否返回高亮片段"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/search/suggestions": {
      "get": {
        "tags": [
          "搜索"
        ],
        "summary": "Search Suggestions",
        "description": "搜索建议（前缀补全），用于搜索框下拉提示\n\n优先使用搜索引擎的建议功能，回退到数据库 LIKE 查询。",
        "operationId": "search_suggestions_api_search_suggestions_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "输入前缀",
              "default": "",
              "title": "Q"
            },
            "description": "输入前缀"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "最大返回数",
              "default": 10,
              "title": "Limit"
            },
            "description": "最大返回数"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/contacts/search": {
      "get": {
        "tags": [
          "联系人"
        ],
        "summary": "Search Contacts",
        "description": "FTS搜索联系人（姓名/电话/微信号/公司/职位/邮箱/备注）",
        "operationId": "search_contacts_api_contacts_search_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "搜索关键词",
              "title": "Q"
            },
            "description": "搜索关键词"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "title": "ApiResponse",
        "description": "统一API响应格式"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [
    {
      "url": "https://www.go-aiport.com",
      "description": "生产环境"
    },
    {
      "url": "https://staging.go-aiport.com",
      "description": "预发布环境"
    },
    {
      "url": "http://localhost:7800",
      "description": "本地开发环境"
    }
  ]
}